# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recall.ai API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved:
    get:
      operationId: getSavedArticles
      tags: [library]
      summary: List saved articles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: collection_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, recall_score, credibility_score, title]
            default: created_at
      responses:
        "200":
          description: List of saved articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticlesResponse"
    post:
      operationId: saveArticle
      tags: [library]
      summary: Save an article to library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveArticleBody"
      responses:
        "201":
          description: Article saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticle"
  /saved/{id}:
    get:
      operationId: getSavedArticle
      tags: [library]
      summary: Get a saved article by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSavedArticle
      tags: [library]
      summary: Delete a saved article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /saved/{id}/collection:
    put:
      operationId: updateArticleCollection
      tags: [library]
      summary: Assign article to a collection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionBody"
      responses:
        "200":
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticle"
  /saved/{id}/share:
    post:
      operationId: shareArticle
      tags: [library]
      summary: Generate a public share link for an article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResponse"
...
  /saved/stats:
    get:
      operationId: getLibraryStats
      tags: [library]
      summary: Get library statistics summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Library statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryStats"
  /share/{token}:
    get:
      operationId: getSharedArticle
      tags: [share]
      summary: Get a publicly shared article by token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /review/due:
    get:
      operationId: getDueReviews
      tags: [insights]
      summary: Get articles due for spaced repetition review
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Articles due for review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedArticle"
  /review/{id}/complete:
    post:
      operationId: completeReview
      tags: [insights]
      summary: Mark an article as reviewed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteReviewBody"
      responses:
        "200":
          description: Review completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedArticle"
...
  /payment/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a Razorpay payment order
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
  /payment/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a payment and upgrade user to pro
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro]
        monthlySavesCount:
          type: integer
        savesLimit:
          type: integer
        preferredLanguage:
          type: string
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - plan
        - monthlySavesCount
        - savesLimit
        - preferredLanguage
        - onboardingCompleted
        - createdAt
...
    SavedArticle:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        url:
          type: ["string", "null"]
        title:
          type: string
        verdict:
          type: string
        bullets:
          type: array
          items:
            type: string
        articleText:
          type: ["string", "null"]
        language:
          type: string
        recallScore:
          type: number
        credibilityScore:
          type: number
        credibilityVerdict:
          type: ["string", "null"]
        collectionId:
          type: ["integer", "null"]
        shareToken:
          type: ["string", "null"]
        isPublic:
          type: boolean
        lastReviewedAt:
          type: ["string", "null"]
          format: date-time
        reviewCount:
          type: integer
        sourceType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - verdict
        - bullets
        - language
        - recallScore
        - credibilityScore
        - isPublic
        - reviewCount
        - sourceType
        - createdAt
    SavedArticlesResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/SavedArticle"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - articles
        - total
        - page
        - totalPages
    SaveArticleBody:
      type: object
      properties:
        url:
          type: ["string", "null"]
        title:
          type: string
        verdict:
          type: string
        bullets:
          type: array
          items:
            type: string
        articleText:
          type: ["string", "null"]
        language:
          type: string
        recallScore:
          type: number
        credibilityScore:
          type: number
        credibilityVerdict:
          type: ["string", "null"]
        collectionId:
          type: ["integer", "null"]
        sourceType:
          type: string
      required:
        - title
        - verdict
        - bullets
        - language
        - recallScore
        - credibilityScore
        - sourceType
    UpdateCollectionBody:
      type: object
      properties:
        collectionId:
          type: ["integer", "null"]
      required: []
    ShareResponse:
      type: object
      properties:
        shareToken:
          type: string
        shareUrl:
          type: string
      required:
        - shareToken
        - shareUrl
    LibraryStats:
      type: object
      properties:
        totalSaves:
          type: integer
        monthlySaves:
          type: integer
        savesLimit:
          type: integer
        avgRecallScore:
          type: number
        avgCredibilityScore:
          type: number
        topTopics:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              count:
                type: integer
            required:
              - topic
              - count
        recentSaves:
          type: array
          items:
            $ref: "#/components/schemas/SavedArticle"
      required:
        - totalSaves
        - monthlySaves
        - savesLimit
        - avgRecallScore
        - avgCredibilityScore
        - topTopics
        - recentSaves
...
    CompleteReviewBody:
      type: object
      properties:
        remembered:
          type: boolean
      required:
        - remembered
...
    PaymentOrder:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
      required:
        - orderId
        - amount
        - currency
        - keyId
    VerifyPaymentBody:
      type: object
      properties:
        orderId:
          type: string
        paymentId:
          type: string
        signature:
          type: string
      required:
        - orderId
        - paymentId
        - signature