# lib/api-zod/src/generated/types/completeReviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recall.ai API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteReviewBody {
  remembered: boolean;
}