# lib/api-zod/src/generated/types/saveArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recall.ai API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveArticleBody {
  /** @nullable */
  url?: string | null;
  title: string;
  verdict: string;
  bullets: string[];
  /** @nullable */
  articleText?: string | null;
  language: string;
  recallScore: number;
  credibilityScore: number;
  /** @nullable */
  credibilityVerdict?: string | null;
  /** @nullable */
  collectionId?: number | null;
  sourceType: string;
}